fix: allow to navigate relative to the current activated route (#5)
Despite workbench uses vanilla Angular router for navigation, relative navigation does not work out of the box.

Analysis:
As of Angular 6.x, commands which target a named outlet (auxiliary route) are not normalized, meaning that
relative navigational symbols like `/`, `./`, or `../` are not resolved (see `create_url_tree.ts` method: `computeNavigation`).
Example which does not resolve relative navigational symbols: `router.navigate([{outlets: {[outlet]: commands}}])`

Fix:
To bypass that restriction, we first create an URL tree without specifying the target outlet. As expected, this translates into an
URL with all navigational symbols resolved. Then, we extract the URL segments of the resolved route and convert it back into commands.
The resulting commands are in their absolute form and may be used for the effective navigation to target a named router outlet.

fixes #5