feat: Allow lazily-loaded views to inject masked injection tokens
Current behavior:
If providing a service in a lazily loaded module with the same injection token as a service provided in the root module,
the service of the root injector is injected into the lazily loaded view instead of the service of the child injector.

Expected behavior:
The service provided in the lazily loaded module should be preferred over the service provided in the root module.