Add iframes of remote sites beyond workbench grid to not cover other parts of the workbench like sashes or view dropdown menu (#30)
Iframes of remote sites are added to a top-level workbench DOM element so they are not reparented upon a workbench layout change. This would cause the iframe to reload. To not cover other parts of the workbench they are placed upfront in the document flow (which makes no z-index necessary). To still receive pointer events, pointer events are disabled for remote site containers (view, activity).

Projection of iframes into their host container's boundary is simplified and positioning properties are set directly via Renderer.