feat: allow cross-origin communication with remote sites
Communication is based on 'postMessage()' and 'onmessage' to safely communicate cross-origin.

Closes #31

BREAKING CHANGE:

Removed output property to listen for URL changes because not allowed for cross-origin communication and internally using a timer to detect URL changes (as there is no change event emitted natively and `MutationObserver` is not applicable). Use `message` output property instead.