Use a separate routing navigate command when closing multiple views all at once (#34)
Otherwise, if a view would reject closing, no view would be closed at all. Also, closing must be done sequentially to have a proper grid snapshot.

In the context of this commit view destruction was consolidated:
- allow to close present views via `WorkbenchRouter.navigate(...)`method by setting `closeIfPresent` in `WbNavigationExtras`
- deprecated property `tryActivateView` and replaced with `activateIfPresent` (consistent naming)
- deprecated method `WorkbenchRouter.resolve()`; the purpose of this method was to identify present views in order to close them; this is now done with `WorkbenchRouter.navigate(...)` method
- moved view destruction from `WorkbenchViewPartService` to `WorkbenchService`

This commit does not break public API but deprecated methods which will be removed in the next release.