Control if workbench part content is capable of being moved in the DOM (#30)
This commit adds the structural component `ContentAsOverlayComponent` which ensures that its content children are not reparented in the DOM when workbench parts are layouted. Technically, content is added to a top-level workbench DOM element and its content projected into that component's bounding box.

For instance, an iframe would reload once it is reparented in the DOM.

Breaking change:
- removed content projection from `RemoteSiteComponent` and added it to workbench part level
  If using a remote site, wrap entire part content in a `<wb-content-as-overlay>` element, which causes it to be added to a top-level workbench DOM element and projected into that component's bounding box.
- removed support to use `RemoteSiteComponent` as a routing component because must be a child of `<wb-content-as-overlay>` element