fix: stretch content of <sci-viewport> if it overflows horizontally
If some content horizontally overflows the viewport boundaries, the viewport correctly shows a horizontal scrollbar. However, when scrolling to the right, background color does not extend to the right and child flex items do not stretch.

For that reason, the display of the viewport is changed from 'flex' to 'grid'. By default, this creates a single column layout which by default stretches its content as expected. If using a flex layout, stretching the flex item would not have the desired effect.

fixes #77

BREAKING CHANGE: CSS display property of <sci-viewport> flex container has changed from `flex` (column nowrap) to `grid` (one column).

To migrate:
- if having a single content child which stretches vertically by using `flex: auto`, remove that property
- if having multiple content children with `flex: none`, wrap them inside a separate flex-container