feat: allow giving CSS classes to workbench elements to have stable selectors available in e2e tests
To have stable element selectors available in e2e tests, dedicated CSS classes were added to fundamental workbench parts. They begin with the prefix `e2e` as a hint to the developer to not remove or rename such classes because being public API and used in e2e tests.

Some elements like a view, popup, activity, message box and notification can be given custom CSS classes to facilitate the selection of those elements.

If showing a notification, a timer is started to close it after some time. The timer was changed to run outside of Angular to allow protractor to continue interacting with the browser while showing the notification.

fixes #78

BREAKING CHANGE: Properties of `Activity` and `WbActivityDirective` to set the activity label and CSS class(es) have changed as following:

- label => itemText
- cssClass => itemCssClass