feat: allow adding actions to the viewpart action bar
Viewpart actions are added to the viewpart action bar located to the right of the view tabs. If added in the context of a view, the action is local to the containing viewpart and only shown if the view is active. If not in the context of a viewpart, the action is added to every viewpart instead.

closes #104