fix: change the iframe url without adding an entry to the browser's history
The iframe URL is now set via `Location.replace` instead of modifying the iframe src attribute. Otherwise, if changing the src attribute of an already attached iframe, a history entry would be added to the browser's history, which would compromise the browser's history back functionality.
Removing the iframe from the DOM before setting the src attribute is also not applicable as this always causes the page to load anew, which is not the desired effect if using hash-based routing and navigating within the same site.

fixes #128