fix: observe element dimension changes natively
`SciDimensionDirective` now listens natively to resize events of its host element instead of observing the Angular zone to detect if the size changed. It uses the built-in `ResizeObserver`, if supported by the user agent, or falls back to listening for resize events on a hidden HTML <object> element.

It is now possible to observe dimension changes programmatically by injecting `SciDimensionService`.

fixes #156

BREAKING CHANGE:
- removed 'viewportChange' output property from `<sci-viewport>` component: instead, add the dimension directive to the viewport and/or viewport client, and/or listen for viewport scroll events with 'scroll' output property