feat: dispatch 'mousemove' and 'mouseup' events across document boundaries
Events of other documents are published on the document's event bus as synthetic 'sci-mousemove' and 'sci-mouseup' events.

Mouse event dispatching is important when using custom scrollbars which are positioned at the iframe border. It allows the user
to scroll seamlessly even when the mouse cursor leaves the iframe, which is because by default, mouse events are only received
by the currently hovering document.