fix: show the sashbox splitter even if no height is set for the sashbox
If the sashbox is not given an explicit height or is not placed inside a growinig/shrinking container, then the splitter height computes to 0px even if set to 100%.
This behavior is fixed by setting the splitter cross-axis size to 'auto' by default.

Also added a testing page for the sashbox and fixed some minor issues.