feat: allow intercepting messages before their publication
An interceptor is registered as a multi bean in the bean manager under the symbol 'MessageInterceptor' or 'IntentInterceptor', respectively.
Multiple interceptors are allowed and build an interceptor chain.

An interceptor can decide to continue publishing by passing the message to the next handler in the chain, or to reject publishing by throwing
an error, or to swallow a message by not calling the next handler at all. If rejecting publishing, the error is transported to the message publisher.