fix(workbench): wait to navigate until other navigations complete
Workbench routing is stateful, causing unexpected behavior when there are multiple navigations at the same time,
 e.g., when a navigation has to wait for a microtask or macrotask to complete. For that reason, routing is now handled exclusively by the
the `WorkbenchRouter`, which acts as a sentinel and queues navigation requests if one is already in progress. To still ensure navigations
to operate on the most current layout, we now modify the layout right before navigation.

We also changed the spec for testing the layout to not run in the `fakeAsync` zone to have a more realistic execution environment.