feat(workbench/popup): allow to open a popup from a screen coordinate and bind it to the lifecycle of a view
Preparing for the microfrontend popup integration, we added support to use a screen coordinate as popup anchor and optionally bind it to a view's lifecycle, allowing to switch between views without closing open popups. As part of the change, we consolidated the API to open the popup and provided e2e protractor tests.

BREAKING CHANGE: consolidated the config for opening a popup in preparation for the microfrontend popup integration

To migrate:
- Rename the `position` property to `align`. This property is used for aligning the popup relative to its anchor.
- Remove the closing strategy `onLayoutChange` as binding a popup to a Workbench view is now supported. This strategy existed only as a workaround to close popups when switching between views.
- Pass the preferred popup overlay size as `PopupSize` object literal instead of separate top-level config properties, as follows:
	- `PopupConfig.width` -> `PopupConfig.size.width`
	- `PopupConfig.height`-> `PopupConfig.size.height`
	- `PopupConfig.minWidth` -> `PopupConfig.size.minWidth`
	- `PopupConfig.maxWidth` -> `PopupConfig.size.maxWidth`
	- `PopupConfig.minHeight` -> `PopupConfig.size.minHeight`
	- `PopupConfig.maxHeight`-> `PopupConfig.size.maxHeight`