feat(workbench-client/popup): allow providing a microfrontend for display in a workbench popup
A micro application can provide a microfrontend for display in a workbench popup. Popup microfrontends are contributed in the form of popup capabilities. A popup is a visual workbench component for displaying content above other content.

Unlike views, popups are not part of the persistent workbench navigation, meaning that popups do not survive a page reload.