feat(workbench-client/message-box): allow messages to be displayed from microfrontends
The built-in message box supports the display of a plain text message and is available as 'message-box' capability without a qualifier. Other message box capabilities can be contributed in the host app via {MicrofrontendMessageBoxProvider}, e.g., to display structured content or to provide out-of-the-box message templates. The use of a qualifier distinguishes different message box providers.

Unlike views, message boxes are not part of the persistent workbench navigation, meaning that message boxes do not survive a page reload.