feat(workbench-client/notification): allow notifications to be displayed from microfrontends
The built-in notification supports the display of a plain text message and is available as 'notification' capability without a qualifier. Other notification capabilities can be contributed in the host app via {MicrofrontendNotificationProvider}, e.g., to display structured content or to provide out-of-the-box notification templates. The use of a qualifier distinguishes different notification providers.