feat(workbench/popup): allow the host app to provide popup capabilities
The host app can now define a popup capability analogous to other micro apps and make it publicly available to be opened by microfrontends.

closes #270

BREAKING CHANGE: Adding support for opening a popup of the host app from within a microfrontend introduced a breaking change in the host/client
communication protocol.

The communication protocol between host and client HAS CHANGED for opening a popup. You need to update host and affected clients to the new version simultaneously. The API has not changed; the breaking change applies only to the `@scion/microfrontend-platform` version.