refactor(workbench/view): open views from within an interceptor
BREAKING CHANGE: Opening views from within an interceptor introduced
a breaking change in the host/client communication protocol.

The communication protocol between host and client HAS CHANGED for
opening a view. You need to update host and affected clients to the new
version simultaneously. The API has not changed; the breaking change
applies only to the version of @scion/workbench and
@scion/workbench-client. To migrate, upgrade to
@scion/workbench@12.0.0-beta.2 and @scion/workbench-client@1.0.0-beta.8,
respectively.