chore: upload workbench and workbench-client packages separately after their build
Ivy is disabled for @scion/workbench-client. When building @scion/workbench,
which has Ivy enabled and lists the @scion/workbench-client as peer dependency,
NGCC is run to transform the workbench client library to an Ivy version.
Since it is not possible to publish a library that has been compiled by NGCC,
we have to make sure that the built library is not modified afterwards.