chore(workbench): migrate deprecated component construction in `WbComponentPortal`
Angular 13 has deprecated the `ComponentFactoryResolver`. Angular 14 provides a new API to construct components without having to attach them to the DOM, which was not possible in Angular 13.

However, we can no longer construct views and viewparts in the `WorkbenchUrlObserver`. Otherwise the routing injection context would be incorrect. This was also the case with the deprecated `ComponentFactoryResolver`, but we managed to replace the injection context. However, this behavior was never documented by Angular and consequently is not supported anymore.

In order to still create inactive views, e.g., when loading a layout into the workbench, we moved their creation to the viewpart component, giving us the correct injection context.