feat(workbench/popup): allow positioning of a popup relative to its contextual view or the page viewport
If positioning the popup by coordinate and opening it in the context of a view, the popup is positioned relative to the bounding box of that view. Otherwise, it is positioned relative to the page viewport. If the view or page viewport is moved or resized, the popup is positioned accordingly.

The following coordinate pairs are supported:
- x/y: relative to the "top/left" corner of the view or page viewport
- top/left: equivalent to passing a "x/y" coordinate
- top/right: relative to the "top/right" corner of the view or page viewport
- bottom/left: relative to the "bottom/left" corner of the view or page viewport
- bottom/right: relative to the "bottom/right" corner of the view or page viewport

closes #342