fix(workbench/host): register application-specific messaging interceptors before workbench/platform interceptors
Previously, if the application provided an initializer under the DI token `MICROFRONTEND_PLATFORM_PRE_STARTUP` to register a messaging interceptor, then that interceptor was installed after workbench or platform messaging interceptors.
This was incorrect because for some messages, application-specific interceptors may never have been invoked. For example, the workbench installs an interceptor to consume view intents to open the workbench view, which means that the chain of interceptors is not continued.