feat(workbench/client): enable Observables of `WorkbenchView` to emit in the correct context
This feature is critical for Angular applications to have Observables of `WorkbenchView` emit in the same Angular zone in which they are subscribed. The emission context can be controlled with an `ObservableDecorator`.

For detailed instructions on how to register an `ObservableDecorator` , refer to https://scion-microfrontend-platform-developer-guide.vercel.app/#chapter:angular-integration-guide:synchronizing-rxjs-observables-with-angular-zone.