fix(workbench): display perspective also for slow/asynchronous initial navigation
If the initial navigation is asynchronous, such as when lazy loading components or using asynchronous guards, the activation of the initial perspective may occur at the same time as the initial navigation. This is problematic because the initial navigation applies a "default" grid with only the main area. To prevent the initial navigation from overriding the layout of the initial perspective, we mark the perspective as active after the initial navigation has completed.