feat(workbench/dialog): support custom header and footer
## Dialog Header
By default, the dialog displays the title and a close button in the header. Alternatively, the dialog supports the use of a custom header.

To provide a custom header, add an Angular template to the HTML of the dialog component and decorate it with the `wbDialogHeader` directive.

```html
<ng-template wbDialogHeader>
  <app-dialog-header/>
</ng-template>
```

## Dialog Footer
A dialog has a default footer that displays actions defined in the HTML of the dialog component. An action is an Angular template decorated with
the `wbDialogAction` directive. Multiple actions are supported, rendered in modeling order, and can be left- or right-aligned.

```html
<!-- Checkbox -->
<ng-template wbDialogAction align="start">
  <label>
    <input type="checkbox"/>
    Do not ask me again
  </label>
</ng-template>

<!-- OK Button -->
<ng-template wbDialogAction align="end">
  <button (click)="...">OK</button>
</ng-template>

<!-- Cancel Button -->
<ng-template wbDialogAction align="end">
  <button (click)="...">Cancel</button>
</ng-template>
```

Alternatively, the dialog supports the use of a custom footer. To provide a custom footer, add an Angular template to the HTML of the dialog component and
decorate it with the `wbDialogFooter` directive.

```html
<ng-template wbDialogFooter>
  <app-dialog-footer/>
</ng-template>
```

BREAKING CHANGE: Support for custom header and footer has introduced a breaking change.

To migrate:
- Type of `WorkbenchDialog.padding` is now `boolean`. Set to `false` to remove the padding, or set the CSS variable `--sci-workbench-dialog-padding` for a custom padding.
- `--sci-workbench-dialog-header-divider-color` CSS variable has been removed (no replacement).