fix(workbench/dialog): do not block dialogs of other views
Previously, the glass pane was rendered in the dialog overlay. Since new dialogs are rendered over existing dialogs, it may no longer be possible to interact with previously opened dialogs.

The glass pane is now added to the DOM as a direct child of the element to be blocked. This can be a view, a popup, a dialog, a microfrontend view or the workbench.