feat(workbench-client/dialog): enable microfrontend display in a dialog
A micro application can provide a microfrontend for display in a workbench dialog. Dialog microfrontends are contributed in the form of dialog capabilities. A dialog is a visual element for focused interaction with the user, such as prompting the user for input or confirming actions. Displayed on top of other content, a dialog blocks interaction with other parts of the application.

## Provide a dialog as follows:
  ```json
  {
    "type": "dialog",
    "qualifier": {
  	"some": "dialog"
    },
    "params": [
  	{"name": "id", "required": true},
    ],
    "properties": {
  	"path": "path/to/dialog",
  	"title": "Dialog Title",
  	"size": {
  	  "height": "500px",
  	  "width": "600px"
  	}
    }
  }
  ```

## Open a dialog as follows:
  ```ts
  inject(WorkbenchDialogService).open({some: 'dialog'}, {
    params: {id: '123'},
  });
  ```