feat(workbench-client/dialog): allow providing a microfrontend for display in a workbench dialog
A micro application can provide a microfrontend for display in a workbench dialog. Dialog microfrontends are contributed in the form of dialog capabilities. A dialog is a visual element for focused interaction with the user, such as prompting the user for input or confirming actions. Displayed on top of other content, a dialog blocks interaction with other parts of the application.