chore(workbench/router): remove option to close view via workbench router link
The router link behaves differently depending on whether it is used inside or outside a view and whether the user presses the CTRL key (Mac: ⌘, Windows: ⊞). Closing a view should be an explicit action, regardless of the environment. We have, therefore, removed the option to close a view using the router link.

BREAKING CHANGE: Removed the option to close a view via the `wbRouterLink` directive.

The router link can no longer be used to close a view. To close a view, use the `WorkbenchView`, the `WorkbenchRouter`, or the `WorkbenchService` instead.

Examples:
```ts
// Closing a view via `WorkbenchView` handle
inject(WorkbenchView).close();

// Closing view(s) via `WorkbenchRouter`
inject(WorkbenchRouter).navigate(['path/*/view'], {close: true});

// Closing view(s) via `WorkbenchService`
inject(WorkbenchService).closeViews('view.1', 'view.2');
```