feat(workbench/router): provide API to to modify the workbench layout
The workbench router provides a new API for modifying the workbench layout. Instead of an array of commands, a function can be passed to the `navigate` method. The router will invoke this function with the current layout. The layout has methods for modification, with each modification creating a new layout instance. Use the new instance for further modifications and finally return it.

The following example adds a part to the left of the main area, inserts a view and navigates it.

```ts
import {inject} from '@angular/core';
import {MAIN_AREA, WorkbenchRouter} from '@scion/workbench';

inject(WorkbenchRouter).navigate(layout => layout
  .addPart('left', {relativeTo: MAIN_AREA, align: 'left'})
  .addView('navigator', {partId: 'left'})
  .navigateView('navigator', ['path/to/view'])
  .activateView('navigator')
);
```