fix(workbench/perspective): create default perspective if no perspective exists
Previously, the workbench created the default perspective if no perspective was configured and not if no perspective was registered. For example, an application can register perspectives directly via the `WorkbenchService` and not the `WorkbenchConfiguration`.

BREAKING CHANGE: The return type of the function to select the initial perspective has changed. To migrate, return the perspective id instead of the perspective instance.