fix(workbench): fix issues related to child routes with an empty-path parent route
The workbench creates an empty-path auxiliary route for each view and part and adds the primary routes defined in the router config as child routes.

The empty-path auxiliary routes caused the following issues for applications having an empty-path parent route:
- Redirecting in a `CanActivate` guard on an empty-path parent route has led to an infinite loop.
- The "Not Found" page was not displayed for an empty-path navigation or when the URL's outlets were cleared.
- The "Nothing to Show" page was not displayed at all.

The issues are caused because Angular always attempts to match empty-path auxiliary routes if the outlet is not included in the URL, regardless of whether the outlet has been navigated to the empty-path route or not. Additionally, Angular's route matcher does not backtrack if it cannot find a route for an empty-path outlet in an empty-path subtree. This prevents fallback to a top-level wildcard route, such as a "Not Found" or "Nothing to Show" route.

closes #655