refactor(workbench): remove deprecated configuration for displaying start page
BREAKING CHANGE: Removed deprecated configuration for displaying start page.

Previously, the component associated with the empty path route was used as the start page.

To migrate for layouts with a main area:
The main area must now be navigated. Previously, no navigation was required and the component associated with the empty path route was used as the start page.

**Example for navigating the main area:**
```ts
bootstrapApplication(AppComponent, {
  providers: [
    provideWorkbench({
      layout: factory => factory
        .addPart(MAIN_AREA)
        .navigatePart(MAIN_AREA, ['path/to/desktop'])
    }),
    provideRouter([
      {
        path: 'path/to/desktop',
        component: DesktopComponent,
      }
    ])
  ]
});
```

**Example for navigating the main area to the empty path route:**
```ts
bootstrapApplication(AppComponent, {
  providers: [
    provideWorkbench({
      layout: factory => factory
        .addPart(MAIN_AREA)
        .navigatePart(MAIN_AREA, [], {hint: 'desktop'}) // pass hint to match a specific empty path route
    }),
    provideRouter([
      {
        path: '',
        component: DesktopComponent,
        canMatch: [canMatchWorkbenchPart('desktop')] // match only if navigating with the specified hint
      }
    ])
  ],
});
```
To migrate for layouts without a main area:
Provide a desktop using an `<ng-template>` with the `wbDesktop` directive. The template content will be used as the desktop content. Previously, the component associated with the empty path route was used as the start page.

```html
<wb-workbench>
  <ng-template wbDesktop>
    Welcome
  </ng-template>
</wb-workbench>
```